[STAD-501] Extract authentication
