+++ synchronization/src/main/java/de/cyface/synchronization/AuthenticatorService.java
@@ -1,12 +1,15 @@
 package de.cyface.synchronization;
 
-import static de.cyface.synchronization.HttpConnection.TAG;
+import static de.cyface.synchronization.Constants.TAG;
 
 import android.app.Service;
 import android.content.Intent;
 import android.os.IBinder;
+import android.preference.PreferenceManager;
 import android.util.Log;
 import androidx.annotation.NonNull;
+
+import de.cyface.uploader.DefaultAuthenticator;
 
 /**
  * The Android service used to communicate with the Stub Authenticator. This has been implemented as described in
@@ -15,7 +18,8 @@ import androidx.annotation.NonNull;
  * Android documentation</a>.
  *
  * @author Klemens Muthmann
- * @version 1.0.4
+ * @author Armin Schnabel
+ * @version 1.0.5
  * @since 2.0.0
  */
 public final class AuthenticatorService extends Service {
@@ -27,7 +31,15 @@ public final class AuthenticatorService extends Service {
     @Override
     public void onCreate() {
         Log.d(TAG, "authenticator service on create!");
-        authenticator = new CyfaceAuthenticator(this);
+
+        // Load authUrl
+        final var preferences = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
+        final var url = preferences.getString(SyncService.SYNC_ENDPOINT_URL_SETTINGS_KEY, null);
+        if (url == null) {
+            throw new IllegalStateException(
+                    "Server url not available. Please set the applications server url preference.");
+        }
+        authenticator = new CyfaceAuthenticator(this, new DefaultAuthenticator(url));
     }
 
     @Override
