+++ synchronization/src/main/java/de/cyface/synchronization/SyncAdapter.kt
@@ -30,7 +30,6 @@ import android.content.SyncResult
 import android.content.pm.PackageManager
 import android.os.Build
 import android.os.Bundle
-import android.preference.PreferenceManager
 import android.util.Log
 import de.cyface.model.RequestMetaData
 import de.cyface.persistence.DefaultPersistenceBehaviour
@@ -41,13 +40,15 @@ import de.cyface.persistence.model.MeasurementStatus
 import de.cyface.persistence.model.ParcelableGeoLocation
 import de.cyface.persistence.serialization.MeasurementSerializer
 import de.cyface.synchronization.ErrorHandler.ErrorCode
+import de.cyface.uploader.Authenticator
+import de.cyface.uploader.Result
+import de.cyface.uploader.UploadProgressListener
+import de.cyface.uploader.Uploader
+import de.cyface.uploader.exception.SynchronizationInterruptedException
 import de.cyface.utils.CursorIsNullException
 import de.cyface.utils.Validate
 import kotlinx.coroutines.runBlocking
 import java.io.File
-import de.cyface.uploader.Result
-import de.cyface.uploader.UploadProgressListener
-import de.cyface.uploader.exception.SynchronizationInterruptedException
 
 /**
  * An Android SyncAdapter implementation to transmit data to a Cyface server.
@@ -57,15 +58,17 @@ import de.cyface.uploader.exception.SynchronizationInterruptedException
  *
  * @author Armin Schnabel
  * @author Klemens Muthmann
- * @version 3.0.2
+ * @version 4.0.0
  * @since 2.0.0
- * @property http The http connection to use for synchronization.
+ * @property authenticator The authenticator to use for synchronization.
+ * @property uploader The uploader to use for synchronization.
  */
 class SyncAdapter private constructor(
     context: Context,
     autoInitialize: Boolean,
     allowParallelSyncs: Boolean,
-    private val http: Http
+    private val authenticator: Authenticator,
+    private val uploader: Uploader
 ) : AbstractThreadedSyncAdapter(context, autoInitialize, allowParallelSyncs) {
 
     private val progressListener: MutableCollection<ConnectionStatusListener>
@@ -81,13 +84,20 @@ class SyncAdapter private constructor(
      *
      * @param context The context this adapter is active under.
      * @param autoInitialize More details are available at `AbstractThreadedSyncAdapter`.
-     * @param http The http connection to use for synchronization.
+     * @param authenticator The authenticator to use for synchronization.
+     * @param uploader The uploader to use for synchronization.
      */
-    internal constructor(context: Context, autoInitialize: Boolean, http: Http) : this(
+    internal constructor(
+        context: Context,
+        autoInitialize: Boolean,
+        authenticator: Authenticator,
+        uploader: Uploader
+    ) : this(
         context,
         autoInitialize,
         false,
-        http
+        authenticator,
+        uploader
     )
 
     init {
@@ -114,7 +124,7 @@ class SyncAdapter private constructor(
                 context,
                 DefaultPersistenceBehaviour()
             )
-        val authenticator = CyfaceAuthenticator(context)
+        val authenticator = CyfaceAuthenticator(context, authenticator)
         val syncPerformer = SyncPerformer(context)
         try {
             // Ensure user is authorized before starting synchronization
@@ -158,7 +168,6 @@ class SyncAdapter private constructor(
 
                     // Acquire new auth token before each synchronization (old one could be expired)
                     val jwtAuthToken = getAuthToken(authenticator, account)
-                    val endPointUrl = getApiUrl(context)
 
                     // Check whether the network settings changed to avoid using metered network without permission
                     if (isSyncRequestAborted(account, authority)) {
@@ -182,7 +191,7 @@ class SyncAdapter private constructor(
                         }
                     }
                     val result = syncPerformer.sendData(
-                        http, syncResult, endPointUrl, metaData,
+                        uploader, syncResult, metaData,
                         compressedTransferTempFile!!, processListener, jwtAuthToken
                     )
                     if (result == Result.UPLOAD_FAILED) {
@@ -216,7 +225,11 @@ class SyncAdapter private constructor(
                     if (compressedTransferTempFile != null && compressedTransferTempFile!!.exists()) {
                         Validate.isTrue(compressedTransferTempFile!!.delete())
                     }
-                    provider.close()
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                        provider.close()
+                    } else {
+                        provider.release()
+                    }
                 }
             }
         } catch (e: CursorIsNullException) {
@@ -278,7 +291,6 @@ class SyncAdapter private constructor(
             Log.w(TAG, "getAuthToken failed, was the connection closed? Aborting sync.")
             throw e
         }
-        @Suppress("FoldInitializerAndIfToElvis") // For readability
         if (bundle == null) {
             // Because of Movebis we don't throw an IllegalStateException if there is no auth token
             throw AuthenticatorException("No valid auth token supplied. Aborting data synchronization!")
@@ -291,22 +303,6 @@ class SyncAdapter private constructor(
         }
         Log.d(TAG, "Login authToken: **" + jwtAuthToken.substring(jwtAuthToken.length - 7))
         return jwtAuthToken
-    }
-
-    /**
-     * Reads the Collector API URL from the preferences.
-     *
-     * @param context The `Context` required to read the preferences
-     * @return The URL as string
-     */
-    private fun getApiUrl(context: Context): String {
-        val preferences = PreferenceManager.getDefaultSharedPreferences(context)
-        val endPointUrl = preferences.getString(SyncService.SYNC_ENDPOINT_URL_SETTINGS_KEY, null)
-        Validate.notNull(
-            endPointUrl,
-            "Sync canceled: Server url not available. Please set the applications server url preference."
-        )
-        return endPointUrl!!
     }
 
     /**
