+++ synchronization/src/main/java/de/cyface/synchronization/SyncService.java
@@ -1,9 +1,17 @@
 package de.cyface.synchronization;
 
 import android.app.Service;
+import android.content.Context;
 import android.content.Intent;
 import android.os.IBinder;
+import android.preference.PreferenceManager;
 import android.util.Log;
+
+import androidx.annotation.NonNull;
+
+import de.cyface.uploader.DefaultAuthenticator;
+import de.cyface.uploader.DefaultUploader;
+import de.cyface.utils.Validate;
 
 /**
  * The synchronisation <code>Service</code> used to bind the synchronisation adapter to the Android framework.
@@ -23,7 +31,7 @@ public final class SyncService extends Service {
      * Logging TAG to identify logs associated with the {@link WiFiSurveyor}.
      */
     @SuppressWarnings({"FieldCanBeLocal", "unused"}) // we add and move logs often, so keep it
-    public static final String TAG = Constants.TAG + ".syncsrvc";
+    public static final String TAG = Constants.TAG + ".syncSvc";
     /**
      * The settings key used to identify the settings storing the URL of the server to upload data to.
      */
@@ -45,7 +53,9 @@ public final class SyncService extends Service {
         Log.v(TAG, "onCreate");
         synchronized (LOCK) {
             if (syncAdapter == null) {
-                syncAdapter = new SyncAdapter(getApplicationContext(), true, new HttpConnection());
+                final var apiEndpoint = apiUrl(getApplicationContext());
+                syncAdapter = new SyncAdapter(getApplicationContext(), true, new DefaultAuthenticator(apiEndpoint),
+                        new DefaultUploader(apiEndpoint));
             }
         }
     }
@@ -54,5 +64,20 @@ public final class SyncService extends Service {
     public IBinder onBind(Intent intent) {
         Log.v(TAG, "onBind");
         return syncAdapter.getSyncAdapterBinder();
+    }
+
+    /**
+     * Reads the Collector API URL from the preferences.
+     *
+     * @param context The `Context` required to read the preferences
+     * @return The URL as string
+     */
+    private String apiUrl(@NonNull final Context context) {
+        final var preferences = PreferenceManager.getDefaultSharedPreferences(context);
+        final var apiEndpoint = preferences.getString(SyncService.SYNC_ENDPOINT_URL_SETTINGS_KEY, null);
+        Validate.notNull(
+                apiEndpoint,
+                "Sync canceled: Server url not available. Please set the applications server url preference.");
+        return apiEndpoint;
     }
 }
